Move apps serving in separate file
